Renamed all headers to correct C++ name styling
