+++ src/see.cpp
@@ -19,9 +19,10 @@
 */
 
 #include <optional>
-#include "board.h"
-#include "bitboard.h"
-#include "types.h"
+
+#include "board.hpp"
+#include "bitboard.hpp"
+#include "types.hpp"
 
 namespace {
