+++ tuner/main.cpp
@@ -26,14 +26,14 @@
 
 #include <fmt/format.h>
 
-#include "tune.h"
-#include "file_resolver.h"
-#include "../src/board.h"
-#include "../src/bitboard.h"
-#include "../src/magic.h"
-#include "../src/transpositional.h"
-#include "../src/tpool.h"
-#include "../cli/cli_parser.h"
+#include "tune.hpp"
+#include "file_resolver.hpp"
+#include "../src/board.hpp"
+#include "../src/bitboard.hpp"
+#include "../src/magic.hpp"
+#include "../src/transpositional.hpp"
+#include "../src/tpool.hpp"
+#include "../cli/cli_parser.hpp"
 
 namespace {
